chore: moved explore and credit-score files
