+++ src/components/features/explore/tabs/all-tab/banner/card-earn-credential.tsx
