+++ src/components/features/explore/tabs/gates-tab/gates-tab.tsx
