+++ src/components/features/explore/tabs/gates-tab/table-view.tsx
