+++ src/pages/explore/earn.tsx
@@ -1,7 +1,7 @@
 import useTranslation from 'next-translate/useTranslation';
 
 import { HeadContainer } from '@/components/molecules/head-container';
-import { GatesTab, ExploreLayout } from '@/components/templates/explore';
+import { GatesTab, ExploreLayout } from '@/components/features/explore';
 
 export default function Earn() {
   const { t } = useTranslation('explore');
