+++ src/pages/explore/index.tsx
@@ -2,7 +2,7 @@ import { InferGetStaticPropsType } from 'next';
 import useTranslation from 'next-translate/useTranslation';
 
 import { HeadContainer } from '@/components/molecules/head-container';
-import { AllTab, ExploreLayout } from '@/components/templates/explore';
+import { AllTab, ExploreLayout } from '@/components/features/explore';
 import { gqlAnonMethods } from '@/services/hasura/api';
 import { useQuery } from '@tanstack/react-query';
