+++ src/pages/model/[id]/recipients.tsx
@@ -2,7 +2,7 @@ import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 
 import { IColumnGrid } from '@/components/organisms/data-grid/data-grid';
-import { DataModelsTab } from '@/components/templates/explore';
+import { DataModelsTab } from '@/components/features/explore';
 import { DataModelLayout } from '@/components/features/protocol/data-models/show/layout';
 import { query } from '@/constants/queries';
